refactor(work): extract transcript markers for quality gate (GH-695)
The branch touched agent-detection.js, which sat on .quality-exceptions
with pre-existing max-lines and cyclomatic-complexity violations — the
gate correctly re-escalated them to errors (allowlisted-but-touched).
Fix by refactoring, not re-allowlisting:

- move extractEntryText / readInitialMarkers / isDispatchedAgentContext
  verbatim to new lib/transcript-markers.js; the marker-window scan
  lives in a hasInitialSubagentMarkers helper (complexity back under
  the cap)
- agent-detection.js requires and re-exports isDispatchedAgentContext,
  so every consumer keeps its import path (323 lines, under the cap)
- remove agent-detection.js from .quality-exceptions — the file now
  passes all six static rules, so the exception is stale (burn-down
  direction: entries only ever removed, never added)

No behavior change; agent-detection + enforce-step-workflow suites and
the GH-695 matrix stay green.
